# [Performance][PhpParser] Improve performance of AstResolver to use existing File Object stmts when current file is equal to target file name to check