--- a/src/PhpParser/AstResolver.php
+++ b/src/PhpParser/AstResolver.php
@@ -81,11 +81,6 @@ final class AstResolver
         $classReflection = $methodReflection->getDeclaringClass();
         $fileName = $classReflection->getFileName();
 
-        // probably native PHP method → un-parseable
-        if ($fileName === null) {
-            return null;
-        }
-
         $nodes = $this->parseFileNameToDecoratedNodes($fileName);
         if ($nodes === []) {
             return null;
@@ -134,10 +129,6 @@ final class AstResolver
     public function resolveFunctionFromFunctionReflection(FunctionReflection $functionReflection): ?Function_
     {
         $fileName = $functionReflection->getFileName();
-        if ($fileName === null) {
-            return null;
-        }
-
         $nodes = $this->parseFileNameToDecoratedNodes($fileName);
         if ($nodes === []) {
             return null;
@@ -219,10 +210,6 @@ final class AstResolver
         $traits = [];
         foreach ($classLikes as $classLike) {
             $fileName = $classLike->getFileName();
-            if ($fileName === null) {
-                continue;
-            }
-
             $nodes = $this->parseFileNameToDecoratedNodes($fileName);
             if ($nodes === []) {
                 continue;
@@ -263,10 +250,6 @@ final class AstResolver
         $classReflection = $phpPropertyReflection->getDeclaringClass();
 
         $fileName = $classReflection->getFileName();
-        if ($fileName === null) {
-            return null;
-        }
-
         $nodes = $this->parseFileNameToDecoratedNodes($fileName);
         if ($nodes === []) {
             return null;
@@ -310,8 +293,13 @@ final class AstResolver
     /**
      * @return Stmt[]
      */
-    public function parseFileNameToDecoratedNodes(string $fileName): array
+    public function parseFileNameToDecoratedNodes(?string $fileName): array
     {
+        // probably native PHP → un-parseable
+        if ($fileName === null) {
+            return [];
+        }
+
         if (isset($this->parsedFileNodes[$fileName])) {
             return $this->parsedFileNodes[$fileName];
         }