--- a/src/PhpParser/AstResolver.php
+++ b/src/PhpParser/AstResolver.php
@@ -37,7 +37,6 @@ use Rector\NodeTypeResolver\NodeTypeResolver;
 use Rector\PhpDocParser\NodeTraverser\SimpleCallableNodeTraverser;
 use Rector\PhpDocParser\PhpParser\SmartPhpParser;
 use Throwable;
-use Webmozart\Assert\Assert;
 
 /**
  * The nodes provided by this resolver is for read-only analysis only!
@@ -310,9 +309,7 @@ final class AstResolver
          */
         $file = $this->currentFileProvider->getFile();
 
-        Assert::isInstanceOf($file, File::class);
-
-        if ($file->getFilePath() === $fileName) {
+        if ($file instanceof File && $file->getFilePath() === $fileName) {
             return $this->parsedFileNodes[$fileName] = $file->getNewStmts();
         }