No need to re-read file to reduce IO when possible.