@samsonasik 2023-09-30
@TomasVotruba @staabm tested running rector on rector-src, it seems faster 👍 

**Before**

```
➜  rector-src git:(main) time bin/rector process --clear-cache
 2088/2088 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        

bin/rector process --clear-cache  216.85s user 8.59s system 739% cpu 30.478 total

```

**After**

```
➜  rector-src git:(main) git checkout ast-resolver-performance 
Switched to branch 'ast-resolver-performance'
➜  rector-src git:(ast-resolver-performance) time bin/rector process --clear-cache
 2088/2088 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        

bin/rector process --clear-cache  201.91s user 8.48s system 732% cpu 28.740 total
```