@samsonasik 2023-09-30
@TomasVotruba @staabm tested running phpunit on rector-src, it seems faster 👍

**Before**

```
Time: 00:34.156, Memory: 409.00 MB

OK (3424 tests, 4922 assertions)
vendor/bin/phpunit  32.15s user 2.24s system 98% cpu 34.808 total
```

**After**

```
Time: 00:33.168, Memory: 403.00 MB

OK (3424 tests, 4922 assertions)
vendor/bin/phpunit  31.66s user 2.05s system 99% cpu 33.799 total
```