@TomasVotruba 2023-09-30
Closing to keep `CurrentFileProvider` usage as low a possible as I plan to remove it.